+++ src/main/java/ru/veselov/websocketroomproject/cache/SubscriptionCache.java
@@ -1,7 +1,5 @@
 package ru.veselov.websocketroomproject.cache;
 
-import ru.veselov.websocketroomproject.event.SubscriptionData;
-
 import java.util.Optional;
 import java.util.Set;
