+++ src/main/java/ru/veselov/websocketroomproject/event/handler/impl/UserDisconnectEventHandlerImpl.java
@@ -3,7 +3,7 @@ package ru.veselov.websocketroomproject.event.handler.impl;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
-import ru.veselov.websocketroomproject.event.SubscriptionData;
+import ru.veselov.websocketroomproject.cache.SubscriptionData;
 import ru.veselov.websocketroomproject.event.handler.UserDisconnectEventHandler;
 import ru.veselov.websocketroomproject.model.ChatUser;
 import ru.veselov.websocketroomproject.service.ChatEventMessageService;
