+++ src/main/java/ru/veselov/websocketroomproject/service/RoomSubscriptionService.java
@@ -1,7 +1,7 @@
 package ru.veselov.websocketroomproject.service;
 
 import org.springframework.stereotype.Service;
-import ru.veselov.websocketroomproject.event.SubscriptionData;
+import ru.veselov.websocketroomproject.cache.SubscriptionData;
 
 import java.util.Optional;
 import java.util.Set;
